Enable AppVeyor and Travis CI
Add appveyor.yml and .travis.yml to configure CI servers, skipping unit tests
on Travis CI for the time being. Due to a bug in Mono, we can't currently run
xUnit 2.0 tests on Mono.

Fix ValidationResources.resx by correcting the binary file names. Those file
names did not have the correct case in the .resx file, leading to an error
on Travis CI.

Fix OpenXmlPackage.cs by replacing a constant with a string. This was necessary
due to a weird error message on Travis CI.

Add *.DotSettings files to .gitignore.

Add xunit.runner.console as a solution-level NuGet package. Add nuget.exe to
.nuget folder as a convenient way to restore packages on the command line.
Edit .gitignore to reflect the recommended settings for NuGet.

Add build.open-xml-sdk.cmd to provide a way to build and run unit tests on
the command line.