Fix binary formatter to deserialize internal types correctly
The binary file was created with a 2.6.0 version of the library with no strong name. If Validate is called without this fix, there is a deserialization exception that will cause a crash. By redirecting the binding, we can force the deserialization to load internal types from the current instance of the assembly and don't need to worry about any later version causing a binding failure.