updated PartUriHelper.GetNextSequenceNumber
Added a new list to the PartUriHelper class that contains all of the
contentTypes that should have a 1 appended to the name of the first item
in the package. Updated the GetNextSequenceNumber method to check the
type and return a value of "1" instead of a blank string. This addresses
issue #123.