Code change for issue #169
Reverted back to previous behavior where the number that is returned to
be appended to the part name is in base10 format. There was no apparent
reason why it was changed to be in hex format. In addition, if the
numbers were in Hex format it caused a problem similar to issue #123.