Make simple types comparable and equatable
In conjunction with #485, this commit introduces two base classes,
OpenXmlComparableReferenceType and OpenXmlComparableValueType, which
implement the IComparable, IComparable<T>, and IEquatable<T>
interfaces. With the exception of EnumValue and ListValue, all children
or descendants of OpenXmlSimpleType are changed to become subclasses of
one of those two base classes.

This commit further adds unit tests for the new behavior. The core
tests are contained in the OpenXmlComparableReferenceTypeTests and
OpenXmlComparableValueTypeTests base classes. For each concrete simple
type, subclasses of those base classes create the concrete values to be
used in the generic unit tests.

Lastly, this commit fixes existing unit tests within the SimpleTypeTest
and OpenXmlSimpleValueTest classes, because those unit tests asserted
that two instances with same value but constructed separately are NOT
equal.

Note that the implementation is not yet complete as it lacks comments
for some public methods, for example. We might also have to add further
unit tests.