Correct constraint registry initialization (#502)
The generator had been doing bit field checking for the format versions,
but with a custom enum where Office2007==0. Thus, everything always
included Office2007, even those items that shouldn't. This change is
mirrored in the generator by using FileFormatVersions itself so that 0
is a None value rather than a specific version.