Update semantic constraints to implement IValidator
This simplifies the validation infrastructure and allows us to move toward having single-pass validation in some cases. This change updates SemanticConstraint to implement IValidator and handle the logic required to fit that interface. Semantic validation is now down as part of the schema validation pass (ie reduces a full pass that had been done on the document).