Add a relationship rewriter to help sanitize malformed URIs
Since .NET 4.5, there has been issues using the SDK with relationships that have malformed URIs. Specifically, this seems to be seen when docs have hyperlinks that are not valid URIs. This change adds a handler to rewrite this upon opening a document in case they exist. Subsequent user code can handle any other manipulation that is needed (such as removal of the offending hyperlink).