Start implementing Linq to XML support
This commit adds the first pieces required for Linq to XML support.
Those pieces are not complete or the final word on how we should integrate
this. There are alternatives to providing the LinqExtensions, for example. The
namespace-related classes (W, W10, and W14) are only a small subset of the
classes required to fully cover all schemas. They only serve as an example to
show what those classes would look like.