Multi-target and enhance
This commit changes the code generation project to target whatever the samples target.
At the moment, this is net46, netcoreapp2.1, netcoreapp3.1, and net5.0. Further, the
Main method now takes one optional argument for the target directory into which the
generated classes are written. If no argument is provided, a default is used.