Add lifetime features and example additional feature.
This adds a couple of features that require changes to the internal
implementation of the parts/packages to provide notifications of
changes such as parts being added/removed or the package being
closed. An additional feature that is implemented in a separate
assembly allows to register items that need to be disposed when the
feature is done which builds on the lifecycle features added in the
main assembly.