Build paragraph id on top of features (#1027)
* Build paragraph id on top of features

This change builds the paragraph id generation on top of the features infrastructure. This allows things to be more streamlined and allow the following features:

- Automatically add paragraph ids to elements as they're inserted into the tree
- Replace duplicated paragraph ids with a fresh unique one
- Provide easy extensibility of system via features

As part of this, a new feature is added internally that allows tracking of elements being inserted/removed. This builds on top of the element context, which should be swapped as the element context incurs a lot of allocations while the new events use structus.

This change also moves it out of the core library and into the features library.