Use full qualified name in semantic validation (#1084)
This may lead to some prefixes that are not correct, but it matches the
actual schematron items we use. This had been cleaned up on the backend,
but required loading the SDK file itself in order to do so. By breaking
that dependency, we must now do the same logic. This is being done in
the TryFindAttribute method and does the following:

- Check if there is an exact match. Returns that if found
- Find a match by name only as a back up