Create separate features for staticly typed classes
This is an initial attempt at separating out the infrastructure for
statically typed classes from the framework. This updates things to
start explicitly depending on the generated classes if they are needed.
Over time, we should be able to separate out the generated classes into
their own assembly and keep a "framework" of OpenXml related
functionality.