Implement initial IPackageFeature.Reload functionality (#1303)
This change implements IPackageFeature.Reload for Stream/File based packages. It does this by extending the initial IPackageFeature implementation with a replaceable System.IO.Packaging.Package instance. When it is replaced, the parts will be updated - to the end user, the same IPackagePart will be used and the underlying storage will be swapped out without them knowing or caring.

This also adds a IPackageFeature.Capabilities property that allows a user to query what capabilities a package feature exposes. This is handy to know if it can save, for instance. As part of this, the OpenXmlPackage.CanSave property will become an instance method, because it is dependent on the package feature.

Fixes #1300