Add better support for malformed URIs
An ability to rewrite relationship URIs was introduced in v2.12. However, this required someone to opt in and manually write support. This change attempts to enable support for these malformed URIs out of the box. Later changes will move this to be an opt-in feature due to the potential performance hit (especially for readonly packages that will essentially be duplicated).