Consolidate generator code to aid caching
The incremental generator pipeline can cache things that don't change, but this requires anything that wants to participate to be a part of the same generator. This moves all the generation that requires the data files into a shared generator so that this can be done. One fall out of this is that the files that are generated are all in the same directory.