Remove EnumString and introduce IEnumValue and IEnumFactory<> (#1397)
This reworks how we're doing enum parsing so we are no longer using C# enums. That incurred a cost while parsing and writing and was challenging for AOT and trimming due to its use of reflection. This now uses a struct that implements two interfaces that allows us to have almost free parsing/writing. The only thing here that costs is checking if a value is valid or potentially getting the version.