Add builder infrastructure to streamline package creation (#1474)
This is a work in progress to eanble a middleware pipeline to create
packages. Main benefits of this are:

- Middleware style pipeline to abstract away things required to do to
initialize a package.
- Can simplify the creation of templates from any source (not fully
working atm)
- Enable Clone() functionality to initialize packages in identical ways
- Provide special builders for more AOT-friendly scenarios
- For example, packages currently have all parts registered to them.
Using these builders, we can track what are the required parts and use
that to create specialized builders (see `DebugPackageBuilderExtensions`
as an example)

For now, it will remain internal, but hopefully for v3.1 can be made
public with some support to better enable AOT scenarios.