Enable package initialization that doesn't get copied
With the pipelining added in 3.0.1 used internally for creating packages, it caused any initializing of the package (i.e. from FlatOpc) to be repeated on clones which we don't want. This fixes it for the near term to unblock users, but will be followed with a consolidation of the uses of IPackageInitialzier.