Enable package initialization that doesn't get copied (#1825)
With the pipelining added in 3.0.1 used internally for creating
packages, it caused any initializing of the package (i.e. from FlatOpc)
to be repeated on clones which we don't want. This fixes the templates
(both FlatOpc and document templates) from not being captures in the
package creation but kept separate for an initialization step.

Fixes #1814